requirements.txt
python-telegram-bot==21.6
ccxt==4.3.92
requests==2.31.0